// app/(dashboard)/certificates/electrical-installation-condition-report/[id]/update/installation-details/schema.ts
import { z } from "zod";

export const Schema = z.object({
  maximumDemand: z.string(),
  distributorsFacility: z.boolean(),
  installationEarthElectrodes: z.boolean(),
  earthElectrodeType: z.string(),
  earthElectrodeLocation: z.string(),
  electrodeResistanceToEarth: z.string(),
  earthingConductorMaterial: z.string(),
  earthingConductorCSA: z.string(),
  earthingConductorVerified: z.boolean(),
  mainProtectiveBondingConductorMaterial: z.string(),
  mainProtectiveBondingConductorCSA: z.string(),
  mainProtectiveBondingConductorVerified: z.boolean(),
  waterInstallationPipes: z.string(),
  gasInstallationPipes: z.string(),
  structuralSteel: z.string(),
  oilInstallationPipes: z.string(),
  lightningProtection: z.string(),
  other: z.string(),
  mainSwitchImage: z.string(),
  mainSwitchLocation: z.string(),
  mainSwitchBSNumber: z.string(),
  mainSwitchType: z.string(),
  mainSwitchRating: z.string(),
  mainSwitchPoles: z.string(),
  mainSwitchCurrentRating: z.string(),
  mainSwitchVoltageRating: z.string(),
  mainSwitchRCDOperatingCurrent: z.string(),
  mainSwitchRCDType: z.string(),
  mainSwitchRCDRatedTimeDelay: z.string(),
  mainSwitchRCDMeasuredOperatingTime: z.string(),
});